model-api/requirements.txt
fastapi>=0.110
uvicorn>=0.27
joblib>=1.3
scikit-learn>=1.3
psycopg2-binary>=2.9
pandas>=2.0
sqlalchemy>=2.0
